Relocate `recordConfig` logic and associated methods from `options.go` to `config.go` for better organization; no functional changes.
